requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.3.1
scikit-image>=0.18.3
matplotlib>=3.4.3
tqdm>=4.62.3
flask>=2.0.1
flask-socketio>=5.1.1
opencv-python>=4.5.3
nibabel>=3.2.1
pydicom>=2.2.2
pandas>=1.3.3
SimpleITK>=2.1.1
